changed more duplicates
