+++ clinic/src/app/app.module.ts
@@ -4,18 +4,18 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './pages/home/home.component';
 import { ClientRegistrationComponent } from './client-registration/client-registration.component';
 import { FooterComponent } from './shared/footer/footer.component';
 import { HeaderComponent } from './shared/header/header.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CategoryComponent } from './category/category.component';
 import { RegistrationComponent } from './registration/registration.component';
-import { ClientPageComponent } from './client-page/client-page.component';
-import { AdminPageComponent } from './admin-page/admin-page.component';
-import { DoctorPageComponent } from './doctor-page/doctor-page.component';
+import { ClientPageComponent } from './pages/client-page/client-page.component';
+import { AdminPageComponent } from './pages/admin-page/admin-page.component';
+import { DoctorPageComponent } from './pages/doctor-page/doctor-page.component';
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { BookingPageComponent } from './booking-page/booking-page.component';
+import { BookingPageComponent } from './pages/booking-page/booking-page.component';
 import { AuthInterceptor } from './core/auth/auth.interceptor';
 import { AuthGuard } from './core/auth/auth.guard';
 import { AuthserviceService } from './core/auth/authservice.service';
@@ -28,6 +28,8 @@ import { DropdownModule } from 'primeng/dropdown';
 import { RatingModule } from 'primeng/rating';
 import { CategoryFieldComponent } from './category-field/category-field.component';
 import { CustomTableComponent } from './custom-table/custom-table.component';
+import { BasePageComponent } from './pages/base-page/base-page.component';
+import { PasswordChangeModalComponent } from './password-change-modal/password-change-modal.component';
 
 
 
@@ -47,7 +49,9 @@ import { CustomTableComponent } from './custom-table/custom-table.component';
     BookingPageComponent,
     CardComponent,
     CategoryFieldComponent,
-    CustomTableComponent
+    CustomTableComponent,
+    BasePageComponent,
+    PasswordChangeModalComponent
     ],
   imports: [
     BrowserModule,
