+++ clinic/src/app/pages/admin-page/admin-page.component.css
@@ -388,200 +388,3 @@ span .year {
 }
 
 
-.custom-table {
-  width: 830px;
-  min-height: 587px;
-  border-radius: 10px;
-  background: #FFF;
-  box-shadow: 0px 4px 30px 0px rgba(5, 51, 84, 0.05);
-  }
-
-  @media (max-width: 1622px) {
-    .calendar {
-      width: 500px;
-      align-items: center;
-    }
-    .custom-table {
-      width: 800px;
-    }
-
-    .tdclick {
-      width: 50px;
-     } 
-    
-  }
-
-  @media (max-width: 880px) {
-    .custom-table {
-      width: 500px;
-    }
-    .tdclick {
-      width: 30px;
-    }
-
-    .rowhead {
-      width: 60px;
-     }
-
-     .card-columns {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-     }
-
-     .left-section {
-      width: 400px;
-      align-items: center;
-    }
-
-     .change-sect {
-      display: none;
-     }
-  }
-  
-  .header {
-    width: 100%;
-    height: 48px;
-    background-color: #f2f2f2;
-    gap: 20px;
-    border-right: #000;
-    text-align: center;
-
-  }
-
- .days-botton {
-  display: flex;
-  height: 48px;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
- }
-
-.days-botton span {
-  width: 21px;
-  height: 21px;
-  border-radius: 100%;
-  background-color: #FFF;
-}
-
-.days-botton img {
-  display: flex;
-  margin: 5px 5px 2px 5px;
-}
-
-.img-2 {
-  transform: rotate(180deg);
-}
-  
-.days {
-  display: flex;
-  gap: 5px;
-  text-align: center;
-  align-items: center; 
-  justify-content: center;
-}
-
-.days:last-child {
-  color: #053354;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.day-number {
-  color: #18A4E1;
-  text-align: center;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-table {
-  border-collapse: collapse;
-  width: 100%;
-}
-
-tr {
-  align-self: center;
-}
-
-td {
-  border: 1px solid #ddd;
-  text-align: center;
-  padding: 0px;
-  height: 60px;
-  width: 100px;
-  position: relative;
-}
-
-@media (max-width: 1622px) {
-  /* Style the first .td element inside each table cell */
- 
-}
-
-.activated-text {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #DAFAEE;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-}
-
-.activated-text p {
-  font-weight: bold;
-  color: #3ACF99;
-  text-align: center;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  word-wrap: break-word;
-}
-
-.delete-button {
-  position: absolute;
-  width: 18px;
-  height: 18px;
-  background-color: #FFFFFF;
-  top: 4px;
-  right: 4px;
-  border-radius: 50%;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.delete-button img {
-  width: 6px; /* Adjust the width of the image */
-  height: 6px; /* Adjust the height of the image */
-  display: block;
-}
-
-/* Set the width of the first column */
-td:first-child {
-  width: 130px;
-}
-
-/* Set the background color for the last two columns */
-td:nth-last-child(1),
-td:nth-last-child(2) {
-  background: #FFFFF5;
-
-}
-.custom-table tbody td.activated {
-  background-color: #DAFAEE;
-  border-radius: 5px;
-
-}
-.custom-table tbody td.disactivated {
-  background-color: #FF93A6;
-  border-radius: 5px;
-}
