+++ clinic/src/app/pages/admin-page/admin-page.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { doctorregisteration } from '../core/auth/useregisteration';
-import { AuthserviceService } from '../core/auth/authservice.service';
+import { doctorregisteration } from '../../core/auth/useregisteration';
+import { AuthserviceService } from '../../core/auth/authservice.service';
 import { data } from 'jquery';
 import { MessageService } from 'primeng/api';
+import { TableDataService } from 'src/app/core/auth/table-data-service.service';
+import { BasePageComponent } from '../base-page/base-page.component';
 declare var $: any;
 
 @Component({
@@ -11,33 +13,18 @@ declare var $: any;
   templateUrl: './admin-page.component.html',
   styleUrls: ['./admin-page.component.css']
 })
-export class AdminPageComponent {
+export class AdminPageComponent extends BasePageComponent implements OnInit {
   activeTab: string = 'doctors';
   activeRole: string = 'doctor';
-  tableData: { cols: { value: string; activated: boolean }[] }[] = [];
- 
-  tableHeaders: { num: number; day: string }[] = [
-    { num: 17, day: 'ორშ' },
-    { num: 18, day: 'სამ' },
-    { num: 19, day: 'ოთხ' },
-    { num: 20, day: 'ხუთ' },
-    { num: 21, day: 'პარ' },
-    { num: 22, day: 'შაბ' },
-    { num: 23, day: 'კვი' },
-  ];
 
   randomDoctor: doctorregisteration | undefined;
-  unauthorizedMessageShown: boolean = false;
-  messageToDoctor: boolean = false;
-  doctor?: doctorregisteration ;
-  doctorId: number | undefined; 
-  appointmentCount: number = 0;
-  doctorFirstName: string='';
-  doctorIdNumber: string='';
+
   starIcons: string[] | undefined;
 
 
-  constructor(private router: Router,public authservice:AuthserviceService, private route: ActivatedRoute, private messageService: MessageService) {}
+  constructor(private router: Router,public authservice:AuthserviceService, private route: ActivatedRoute, private messageService: MessageService, public tableDataService: TableDataService) {
+    super();
+  }
   
   ngOnInit() {
 
@@ -121,26 +108,8 @@ export class AdminPageComponent {
       );
     };
     
-
-    for (let i = 1; i <= 9; i++) {
-      const row = { cols: [] as { value: string; activated: boolean }[] };
-      for (let j = 1; j <= 7; j++) {
-        row.cols.push({ value: `${i}-${j}`, activated: false });
-      }
-      this.tableData.push(row);
-    }    
   }
 
-  getTimeRange(rowNumber: number): string {
-    const startTime = 9;
-    const endTime = 18;
-    const timeSlot = 1;
-
-    const startHour = startTime + rowNumber * timeSlot;
-    const endHour = startHour + timeSlot;
-
-    return `${startHour}:00 - ${endHour}:00`;
-  }
 
   setActiveTab(tab: string) {
     this.activeTab = tab;
