+++ clinic/src/app/pages/base-page/base-page.component.html
@@ -1,0 +1,1 @@
+<p>base-page works!</p>
