+++ clinic/src/app/pages/booking-page/booking-page.component.html
@@ -15,14 +15,14 @@
                         
                         <div class="doctors-stars" *ngIf="this.doctor">
                           <ng-container *ngFor="let star of getStarArray(this.doctor?.starNum || 0)">
-                            <img src="../../assets/material-symbols_star-rounded.png" alt="Star">
+                            <img src="../../../assets/material-symbols_star-rounded.png" alt="Star">
                           </ng-container>
                         </div>
                         
                         
                     </div>
                   </div>
-                  <img src="../../assets/Vector 7.png" alt="">
+                  <img src="../../../assets/Vector 7.png" alt="">
                   <div class="cv-style">
                     <p><span class="blue-text">2017 - დღემდე</span>, ჩვენი კლინიკის გენერალური დირექტორი</p>
                     <p><span class="blue-text">2002 - დღემდე</span>, ჩვენი კომპიუტერული ტომოგრაფიის განყოფილების ხელმძღვანელი</p>
@@ -36,9 +36,9 @@
               <div class="time">
                 <span class="year">2023</span>
                 <div class="buttons-click">
-                    <span class="arrow arrow-left "><img src="../../assets/material-symbols_arrow-back-ios-rounded.png" alt=""></span>
+                    <span class="arrow arrow-left "><img src="../../../assets/material-symbols_arrow-back-ios-rounded.png" alt=""></span>
                     <button class="action-button">აპრილი</button>
-                    <span class="arrow arrow-right"><img src="../../assets/material-symbols_arrow-back-ios-rounded (1).png" alt=""></span>
+                    <span class="arrow arrow-right"><img src="../../../assets/material-symbols_arrow-back-ios-rounded (1).png" alt=""></span>
                 </div>
             </div>
               <div class="time-and-table">
@@ -59,11 +59,11 @@
                 </div>
               <div class="change-sect">
                   <button class="redact">
-                      <span><img src="../../assets/edit-svgrepo-com 2.png" alt=""></span>
+                      <span><img src="../../../assets/edit-svgrepo-com 2.png" alt=""></span>
                       <p>რედაქტირება</p>
                   </button>
                 <button class="delete">
-                    <span><img src="../../assets/delete-file-svgrepo-com 3.png" alt=""></span>
+                    <span><img src="../../../assets/delete-file-svgrepo-com 3.png" alt=""></span>
                     <p>წაშლა</p>
                 </button>
             </div>
