+++ clinic/src/app/pages/doctor-page/doctor-page.component.css
@@ -212,155 +212,6 @@ span .year {
   line-height: normal;
 }
 
-.custom-table {
-  width: 830px;
-  min-height: 587px;
-  border-radius: 10px;
-  background: #FFF;
-  box-shadow: 0px 4px 30px 0px rgba(5, 51, 84, 0.05);
-  }
-  
-  .header {
-    width: 100%;
-    height: 48px;
-    background-color: #f2f2f2;
-    gap: 20px;
-    border-right: #000;
-    text-align: center;
-
-  }
-
- .days-botton {
-  display: flex;
-  height: 48px;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
- }
-
-.days-botton span {
-  width: 21px;
-  height: 21px;
-  border-radius: 100%;
-  background-color: #FFF;
-}
-
-.days-botton img {
-  display: flex;
-  margin: 5px 5px 2px 5px;
-}
-
-.img-2 {
-  transform: rotate(180deg);
-}
-  
-.days {
-  display: flex;
-  gap: 5px;
-  text-align: center;
-  align-items: center; 
-  justify-content: center;
-}
-
-.days:last-child {
-  color: #053354;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.day-number {
-  color: #18A4E1;
-  text-align: center;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-table {
-  border-collapse: collapse;
-  width: 100%;
-}
-
-tr {
-  align-self: center;
-}
-
-td {
-  border: 1px solid #ddd;
-  text-align: center;
-  padding: 0px;
-  height: 60px;
-  width: 100px;
-  position: relative;
-}
-
-.activated-text {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #DAFAEE;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-}
-.custom-table tbody td.activated {
-  background-color: #b0f6dc;
-}
-.custom-table tbody td.disactivated {
-  background-color: #f33760;
-}
-.tdclick{
-  cursor: pointer;
-}
-.activated-text p {
-  font-weight: bold;
-  color: #3ACF99;
-  text-align: center;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  word-wrap: break-word;
-}
-
-.delete-button {
-  position: absolute;
-  width: 18px;
-  height: 18px;
-  background-color: #FFFFFF;
-  top: 4px;
-  right: 4px;
-  border-radius: 50%;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.delete-button img {
-  width: 6px; /* Adjust the width of the image */
-  height: 6px; /* Adjust the height of the image */
-  display: block;
-}
-
-/* Set the width of the first column */
-td:first-child {
-  width: 130px;
-}
-
-/* Set the background color for the last two columns */
-td:nth-last-child(1),
-td:nth-last-child(2) {
-  background: #FFFFF5;
-
-}
-
   /* changes buttons */
 
   .change-sect {
@@ -508,155 +359,19 @@ td:nth-last-child(2) {
   width: 16px;
   height: 16px;
   flex-shrink: 0;
-  background-image: url('../../assets/material-symbols_arrow-back-ios-rounded333.png');
+  background-image: url('../../../assets/material-symbols_arrow-back-ios-rounded333.png');
 }
 
 .middle button:hover .buttons-span {
-  background-image: url('../../assets/material-symbol_white_right_arrow.png');
+  background-image: url('../../../assets/material-symbol_white_right_arrow.png');
 }
 
 .middle button:hover {
   border-radius: 25px;
   border: 1px solid #18a4e1;
-  background-color: #fff; /* Updated active background color */
+  background-color: #fff;
   color: #18a4e1;
 }
 .middle button:active {
   overflow: hidden;
-}
-
-
-/* modal */
-
-.modal-content {
-  width: 600px;
-  height: 450px;
-  flex-shrink: 0;
-  border-radius: 10px;
-  background: #FFF;
-  padding-top: 66px;
-  padding-left: 113px;
-  padding-right: 111px;
-  padding-bottom: 89px;
-}
-
-@media (max-width: 880px) {
-  .modal-content {
-    width: 250px;
-    height: 260px;
-    padding: 0 ;
-    padding-top: 20px;
-  }
-
-  form .form-group {
-    width: 200px;
-  }
-
-  .modal-title {
-    font-size: 16px !important;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    margin-bottom: 20px;
-  }
-  .modal-header {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  .modal-body {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 0;
-  }
-  .modal-body form {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin-top: 0 !important;
-  }
-}
-
-.modal-header{
-  border: none;
-  padding: 0px;
-}
-
-.modal-header h5 {
-  color: #18A4E1;
-  font-size: 26px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.modal-body {
-  padding: 0px;
-  margin-bottom: 42px;
-}
-
-.modal-body form {
-  margin-top: 44px;
-  display: grid;
-  gap: 20px;
-}
-form p {
-  margin: 0px;
-  padding: 0px;
-}
-
-.form-group {
-  width: 377px;
-  height: 48px;
-  flex-shrink: 0;
-  margin: 0px;
-  
-}
-
-.form-password {
-  margin-bottom: 5px;
-}
-
-.form-group input {
-  color: rgba(5, 51, 84, 0.30);
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  height: 90%;
-  padding: 15px 20px;
-  border-radius: 50px;
-  border: 1px solid #ECF2F8;
-}
-
-
-.forget-password {
-  color: #053354;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  text-align: right;
-  cursor: pointer;
-}
-
-.modal-footer {
-  flex-shrink: 0;
-  border: none;
-  justify-content: left;
-  padding: 0px;
-}
-
-.modal-footer button {
-  width: 180px;
-  height: 40px;
-  background: #053354;
-  color: #FFF;
-  border-radius: 50px;
-
-}
-
-.change-pass {
-  cursor: pointer;
 }
