+++ clinic/src/app/pages/doctor-page/doctor-page.component.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { doctorregisteration } from '../core/auth/useregisteration';
-import { AuthserviceService } from '../core/auth/authservice.service';
+import { doctorregisteration } from '../../core/auth/useregisteration';
+import { AuthserviceService } from '../../core/auth/authservice.service';
 import { data } from 'jquery';
 import { MessageService } from 'primeng/api';
-import { TableDataService } from '../core/auth/table-data-service.service';
+import { TableDataService } from '../../core/auth/table-data-service.service';
+import { BasePageComponent } from '../base-page/base-page.component';
 declare var $: any;
 
 @Component({
@@ -12,29 +13,15 @@ declare var $: any;
   templateUrl: './doctor-page.component.html',
   styleUrls: ['./doctor-page.component.css'],
 })
-export class DoctorPageComponent implements OnInit{
-
-  // tableData: { cols: { value: string; activated: boolean }[] }[] = [];
- 
-
-  unauthorizedMessageShown: boolean = false;
-  messageToDoctor: boolean = false;
-  doctor?: doctorregisteration ;
-  doctorId: number | undefined; 
-  appointmentCount: number = 0;
-  doctorFirstName: string='';
-  doctorIdNumber: string='';
-  text: string = "";
-  clientName: string = "";
-  tooltipBox: boolean = false;
-  clientLastName: string="";
-  clientIdNumber: string="";
-  oldPassword: string="";
-  newPassword: string="";
-  submissionSuccess: boolean = false;
+export class DoctorPageComponent extends BasePageComponent implements OnInit{
 
 
-  constructor(private router: Router,public authservice:AuthserviceService, private route: ActivatedRoute, private messageService: MessageService, public tableDataService: TableDataService) {}
+
+
+
+  constructor(private router: Router,public authservice:AuthserviceService, private route: ActivatedRoute, private messageService: MessageService, public tableDataService: TableDataService) {
+    super();
+  }
   
   ngOnInit() {
     this.route.params.subscribe(params => {
