+++ clinic/src/app/pages/home/home.component.html
@@ -4,7 +4,7 @@
         <div class="rectangle">
             <div class="rectangle-child">
                 <p>აირჩიე კლინიკა რომელსაც შეგიძლია ენდო</p>
-                <button class="button"><img src="../../assets/location.png" alt="">იაკობ ნიკოლაძე №10</button>
+                <button class="button"><img src="../../../assets/location.png" alt="">იაკობ ნიკოლაძე №10</button>
             </div>
         </div>
         <!-- Other content goes here -->
