+++ clinic/src/app/password-change-modal/password-change-modal.component.css
@@ -1,0 +1,132 @@
+/* modal */
+
+.modal-content {
+    width: 600px;
+    height: 450px;
+    flex-shrink: 0;
+    border-radius: 10px;
+    background: #FFF;
+    padding-top: 66px;
+    padding-left: 113px;
+    padding-right: 111px;
+    padding-bottom: 89px;
+  }
+  
+  @media (max-width: 880px) {
+    .modal-content {
+      width: 250px;
+      height: 260px;
+      padding: 0 ;
+      padding-top: 20px;
+    }
+  
+    form .form-group {
+      width: 200px;
+    }
+  
+    .modal-title {
+      font-size: 16px !important;
+      font-style: normal;
+      font-weight: 400;
+      line-height: normal;
+      margin-bottom: 20px;
+    }
+    .modal-header {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+    .modal-body {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      margin-top: 0;
+    }
+    .modal-body form {
+      display: flex;
+      flex-direction: column;
+      gap: 20px;
+      margin-top: 0 !important;
+    }
+  }
+  
+  .modal-header{
+    border: none;
+    padding: 0px;
+  }
+  
+  .modal-header h5 {
+    color: #18A4E1;
+    font-size: 26px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+  }
+  
+  .modal-body {
+    padding: 0px;
+    margin-bottom: 42px;
+  }
+  
+  .modal-body form {
+    margin-top: 44px;
+    display: grid;
+    gap: 20px;
+  }
+  form p {
+    margin: 0px;
+    padding: 0px;
+  }
+  
+  .form-group {
+    width: 377px;
+    height: 48px;
+    flex-shrink: 0;
+    margin: 0px;
+    
+  }
+  
+  .form-password {
+    margin-bottom: 5px;
+  }
+  
+  .form-group input {
+    color: rgba(5, 51, 84, 0.30);
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+    height: 90%;
+    padding: 15px 20px;
+    border-radius: 50px;
+    border: 1px solid #ECF2F8;
+  }
+  
+  
+  .forget-password {
+    color: #053354;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+    text-align: right;
+    cursor: pointer;
+  }
+  
+  .modal-footer {
+    flex-shrink: 0;
+    border: none;
+    justify-content: left;
+    padding: 0px;
+  }
+  
+  .modal-footer button {
+    width: 180px;
+    height: 40px;
+    background: #053354;
+    color: #FFF;
+    border-radius: 50px;
+  
+  }
+  
+
