+++ clinic/src/app/password-change-modal/password-change-modal.component.html
@@ -1,0 +1,24 @@
+<div class="modal fade" id="passwordChangeModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h5 class="modal-title" id="exampleModalLabel">პაროლის შეცვლა</h5>
+        </div>
+        <div class="modal-body">
+          <form (ngSubmit)="onSubmit()">
+            <div class="form-group">
+              <input type="password" class="form-control" name="old-password" id="old-password" placeholder="ძველი პაროლი" [(ngModel)]="oldPassword" required> 
+            </div>
+            <div class="form-group">
+              <input type="password"  name="newPassword" class="form-control form-password" id="recipient-name" placeholder="ახალი პაროლი" [(ngModel)]="newPassword" required> 
+            </div>
+          
+            <div class="modal-footer">
+              <button type="submit" *ngIf="submissionSuccess" data-dismiss="modal">დახურვა</button>
+              <button type="submit" *ngIf="!submissionSuccess">შენახვა</button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </div>
